requirements.txt
slack-bolt
slack-sdk
# openai>=0.28.1,<0.29
tiktoken
# https://github.com/Yelp/elastalert/issues/2306
urllib3
python-dotenv
aiohttp